incorrect embed tags in image index views
https://www.pivotaltracker.com/story/show/171381476

* also includes fix in embeddable_content_seeds.rb for im-preview app

+++ db/embeddable_content/embeddable_content_seeds.rb
@@ -1,9 +1,11 @@
-[Embedding, EmbedderConfig, EmbeddableModelConfig].each do |model|
-  model.destroy_all
-  Pathname.new(File.expand_path('./', __dir__)).tap do |dir|
-    dir.join(model.model_name.plural).sub_ext('.json').tap do |file|
-      JSON.parse(file.read).each do |attrs|
-        model.create attrs
+if Rails.env.test?
+  [Embedding, EmbedderConfig, EmbeddableModelConfig].each do |model|
+    model.destroy_all
+    Pathname.new(File.expand_path('./', __dir__)).tap do |dir|
+      dir.join(model.model_name.plural).sub_ext('.json').tap do |file|
+        JSON.parse(file.read).each do |attrs|
+          model.create attrs
+        end
       end
     end
   end
